backend/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
djangorestframework==3.14.0
easyocr==1.7.1
filelock==3.13.1
fsspec==2024.2.0
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
lazy_loader==0.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
pyclipper==1.3.0.post5
pycodestyle==2.11.1
python-bidi==0.4.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tifffile==2024.2.12
torch==2.2.0
torchvision==0.17.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
